requirements.txt
# Django и DRF
asgiref~=3.8.1
Django~=5.0.8
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.2.0

# Аутентификация и безопасность
djangorestframework-simplejwt>=5.3.0
django-guardian>=2.4.0

# База данных и кэширование
psycopg2-binary>=2.9.9
django-redis>=5.4.0

# Асинхронные задачи
celery>=5.3.6
redis>=5.0.1

# Хранение файлов
django-storages>=1.14.2
boto3>=1.34.14

# AI и ML
torch>=2.1.0
transformers>=4.35.2

# Утилиты
python-magic>=0.4.27
Pillow>=10.1.0
python-dotenv>=1.0.0
python-decouple~=3.8
sqlparse~=0.5.1
django-browser-reload~=1.13.0
watchdog>=3.0.0

# Разработка и тестирование
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.1
pytest>=7.4.3
pytest-django>=4.7.0
pytest-asyncio>=0.21.1
factory-boy>=3.3.0
